HCCGENERAL-129: Fix inability to rotate the QuickPlot capture box. The reason is that the logic to determine what action to take is of the form: if (dragging) { do this } else if (rotating) { do that }
The problem is that in the event that dragging and rotating both evaluate to true (which is probably most of the time) the dragging condition "wins". The fix was to reverse the order of evaluation. Because if it's both true that we're dragging and rotating. The rotating condition should be the one that "wins"